Retrofit, MVVM

+++ android/supplier/DeliveryApp/app/src/main/java/hu/bme/aut/android/deliveryapp/model/JobDetails.kt
@@ -5,7 +5,7 @@ import android.location.Location
 data class JobDetails(
     val name: String,
     val imPaths: String,
-    val avgRating: Int,
+    val avgRating: Float,
     val deliveryDate: String,
     val deliveryCost: String,
     val deliveryLocation: Location
